requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-admin==2.0.2
django-excel-response2==3.0.5
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
idna==3.4
isoweek==1.3.3
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
screen==1.0.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
TimeConvert==3.0.13
typing-extensions==4.6.3
tzlocal==5.0.1
urllib3==2.0.3
xlwt==1.3.0
